Add support for opening FF bundles directly
This involved breaking compatibility with the newer asset bundle versions
that UP supports, as the formats (or assumptions about the formats)
appear to diverge from the ones FF uses.

It's probably desirable to prune the control flow we're not using at
some point, so as to make the codebase more readable, and allow for
further restructuring to support ex. reading multiple asset files from a
single bundle and gracefully ignoring non-asset files.

I'll likely also want to replicate the changes made to
UnityEnvironment.get_asset_by_filename() in the other environment
functions, but it isn't of a high priority.