Fix a few bugs with import_audio
I noticed when trying to create a new audioclip object, using import_audio on it caused exceptions. Likely due to m_Format and m_DecompressOnLoad not being set (format is not a property of the AudioClip class). The size property also didn't seem to be getting set, there is a size property in StreamedResource but I'm not sure if that's used in our code so I replaced it with a more direct _obj.
Tested and works with both replacement and adding new audio.